kapitan-server: in-memory inventory daemon with exact incremental re-rendering
The server holds the rendered inventory, watches the inventory directory
(and the real location of symlinked files), and re-renders only the targets
whose inputs changed. Each target records the files it was rendered from and
every path probed while resolving class names, so creating a shadowing class
file invalidates exactly the right targets. JSON-RPC 2.0 over a unix socket;
the CLI auto-spawns a server per inventory (30 min idle timeout), falls back
to local rendering with --no-daemon, and gains `inventory watch` plus
`server start|stop|status|logs`.

Also: explain moves into the library, resolver arguments carry their kind
(literal / node / computed) so `dict` matches the reference, the emitter no
longer underflows on escaped double-quoted scalars, and a fixture inventory
with reference-generated expected output backs the integration tests.

+++ crates/kapitan-inventory/src/emit/yaml.rs
@@ -729,7 +729,7 @@ impl Emitter {
             if 0 < end
                 && end + 1 < text.len()
                 && (ch == Some(' ') || start >= end)
-                && self.column + (end - start) > self.opts.width
+                && (self.column as i64 + (end as i64 - start as i64)) > self.opts.width as i64
                 && split
             {
                 let mut data: String = if start < end { text[start..end].iter().collect() } else { String::new() };
